feat: initial refactoring
