+++ src/components/subscriptions/MultipleSubscriptionsPage.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const MultipleSubscriptionsPage = () => (
+  <main>
+    <h4>Welcome to Subscription Management!</h4>
+  </main>
+);
+
+export default MultipleSubscriptionsPage;
